fix-pymt-type

+++ db/migrate/20150411044914_create_payments.rb
@@ -2,7 +2,7 @@ class CreatePayments < ActiveRecord::Migration
   def change
     create_table :payments do |t|
       t.string :Pymt_Type
-      t.integer :Pymt_Amt
+      t.float :Pymt_Amt
       t.date :Pymt_Date
       t.references :member, index: true
